requirements.txt
py-cord==1.7.3
aiohttp==3.8.1
PyYAML==6.0
